Refactor large source files into focused modules
