+++ src/altTextConfig.ts
@@ -1,0 +1,2 @@
+export const ALT_TEXT_CHARACTER_LIMIT = 125;
+export const MAX_REWRITE_LOOPS = 10;
