+++ src/styles/controls.css
@@ -1,0 +1,140 @@
+.field {
+  display: grid;
+  gap: 10px;
+  padding: 18px;
+  border-radius: 20px;
+  background: var(--panel-strong);
+  border: 1px solid var(--border);
+}
+
+.warning-card {
+  display: grid;
+  gap: 10px;
+  padding: 18px;
+  border-radius: 20px;
+  background: linear-gradient(180deg, rgba(90, 50, 18, 0.48) 0%, rgba(47, 26, 12, 0.78) 100%);
+  border: 1px solid rgba(243, 139, 49, 0.32);
+  box-shadow: inset 0 0 0 1px rgba(243, 139, 49, 0.08);
+}
+
+.label,
+.status-label,
+.result-label {
+  font-size: 0.86rem;
+  font-weight: 600;
+  letter-spacing: 0.05em;
+  text-transform: uppercase;
+  color: var(--text-muted);
+}
+
+.warning-label {
+  font-size: 0.86rem;
+  font-weight: 700;
+  letter-spacing: 0.05em;
+  text-transform: uppercase;
+  color: #ffd39d;
+}
+
+.warning-body {
+  margin: 0;
+  padding: 14px 16px;
+  border: 1px solid rgba(255, 211, 157, 0.14);
+  border-radius: 16px;
+  background: rgba(18, 11, 6, 0.45);
+  color: #fff2df;
+  white-space: pre-wrap;
+  font-family: 'IBM Plex Mono', 'SFMono-Regular', monospace;
+  font-size: 0.92rem;
+}
+
+.hint {
+  font-size: 0.92rem;
+  color: var(--text-muted);
+}
+
+input[type='file'],
+textarea {
+  width: 100%;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  border-radius: 16px;
+  background: rgba(8, 10, 15, 0.9);
+  color: inherit;
+}
+
+input[type='file'] {
+  padding: 14px;
+}
+
+textarea {
+  min-height: 148px;
+  padding: 14px 16px;
+  resize: vertical;
+}
+
+.prompt-display {
+  margin: 0;
+  padding: 14px 16px;
+  min-height: 148px;
+  border: 1px solid rgba(255, 255, 255, 0.12);
+  border-radius: 16px;
+  background: rgba(8, 10, 15, 0.9);
+  color: inherit;
+  white-space: pre-wrap;
+  font-family: 'IBM Plex Mono', 'SFMono-Regular', monospace;
+  font-size: 0.92rem;
+}
+
+.toggle-field {
+  gap: 12px;
+}
+
+.toggle-row {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  color: inherit;
+}
+
+.toggle-row input[type='checkbox'] {
+  width: 18px;
+  height: 18px;
+  accent-color: var(--accent);
+}
+
+input[type='file']:focus-visible,
+textarea:focus-visible,
+.run-button:focus-visible {
+  outline: 2px solid var(--accent);
+  outline-offset: 2px;
+}
+
+.run-button {
+  position: relative;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  border: 0;
+  border-radius: 999px;
+  padding: 15px 18px;
+  font-weight: 700;
+  color: #15161d;
+  background: linear-gradient(135deg, #f3c560 0%, #f38b31 52%, #ef6a39 100%);
+  box-shadow: 0 18px 34px rgba(239, 106, 57, 0.26);
+}
+
+.run-button[data-busy='true']::before {
+  content: '';
+  width: 16px;
+  height: 16px;
+  border-radius: 50%;
+  border: 2px solid rgba(21, 22, 29, 0.28);
+  border-top-color: #15161d;
+  animation: spin 0.8s linear infinite;
+}
+
+.run-button[disabled] {
+  cursor: wait;
+  opacity: 0.7;
+  box-shadow: none;
+}
